% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMOData.R
\docType{data}
\name{IMOData}
\alias{IMOData}
\title{A samll example dataset}
\format{\describe{
\item
\item
}}
\source{
I made this up
}
\usage{
IMOData
}
\description{
This is just an example to see if I can get
this to work
}
\keyword{datasets}
