% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priemelData.R
\docType{data}
\name{priemelData}
\alias{priemelData}
\title{Another example dataset}
\format{\describe{
 \item
 \item
}}
\source{
I made this up
}
\usage{
priemelData
}
\description{
This is just an example to see if I can get
this to work
}
\keyword{datasets}
