% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sisbidVitD.R
\docType{package}
\name{sisbidVitD}
\alias{sisbidVitD}
\alias{sisbidVitD-package}
\title{A practice package}
\description{
This is just a practice. It includes a function \code{\link{trivial}}.
}
\details{
We can put more things here.
}
