% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivial.R
\name{trivial}
\alias{trivial}
\title{A Trivial Function}
\usage{
trivial(x, y)
}
\arguments{
\item{x, y}{two values to be summed}
}
\value{
the sum x+y
}
\description{
Adding a trivial function
}
\examples{
trivial(2,3)
}
